Split iOS cloud link responsibilities
